+++ CHANGELOG.md
@@ -8,6 +8,15 @@ CHANGELOG for Bikeways Website
 
 ### Version 3.0.0 (03/17/2015)
 
+* added icon for bike safety no earphones
+* added link to ARS documents
+* made changes to onPathsPage text
+* made changes to onStreetPage text
+* made changes to saftyPage text
+* added pdf link to paper map
+* added facebook link
+* added youtube videos
+* added new pics
 * adjusted center point of map
 * added GRID Bike Share Points to the map
 * update to [ArcGIS API for JavaScript 3.13compact] (https://developers.arcgis.com/javascript/index.html)
@@ -74,6 +83,10 @@ CHANGELOG for Bikeways Website
 
 ## Resources
 
+* http://www.azbikeped.org/index.asp
+* http://www.azleg.gov/ArizonaRevisedStatutes.asp?Title=28
+* http://www.azmag.gov/Documents/MAG_2012-10-19_Bike-Pathways-Map.pdf
+
 #### ***Reference Example Sites***
 
     * http://www.bicyclela.org/fullscreenmap.html
@@ -88,6 +101,7 @@ CHANGELOG for Bikeways Website
     * http://36cs.maps.arcgis.com/apps/OnePane/basicviewer/index.html?appid=60dbc11f3f214506b85fb3d9ecd91034
     * http://web6.seattle.gov/SDOT/BikeMap/
     * http://cyclodeo.com/#/
+    * http://gis.drcog.org/bikeroutes/
 
 #### ***Bike Shop Data Base App Reference***